+++ src/Outkeep.Grains.Interfaces/ICacheGrain.cs
@@ -1,5 +1,6 @@
 ﻿using Orleans;
 using Orleans.Concurrency;
+using Outkeep.Grains.Caching;
 using System;
 using System.Threading.Tasks;
