+++ src/Outkeep.Grains/CacheGrain.cs
@@ -1,6 +1,7 @@
 ﻿using Orleans;
 using Orleans.Concurrency;
 using Orleans.Runtime;
+using Outkeep.Grains.Caching;
 using Outkeep.Grains.Governance;
 using System;
 using System.Threading.Tasks;
