+++ test/Outkeep.Api.Http.Tests/CacheControllerV1Tests.cs
@@ -5,6 +5,7 @@ using Orleans;
 using Orleans.Concurrency;
 using Outkeep.Api.Http.Controllers.V1;
 using Outkeep.Grains;
+using Outkeep.Grains.Caching;
 using System;
 using System.IO;
 using System.Linq;
